Feat(layout): layout of the main page
Bulk of layout included. Changed some CSS to move to a darker theme - not all CSS properties are changed, we will change them as we go and need. 

Some basic website skeleton with place for extra pages and resources.

Technologies overview layout currently the same as the default for this theme - this will be changed in the future.